backend-function/requirements.txt
azure-cosmos
azure-functions